Added some example scripts
These are used to generate the results. The script "for the report" is used to generate the final results.
The file 'testing_trainlm_2layers.m' is an example of a script which is used to find the best combination of learning rates, regularization values, and the number of neurons and layers.